trade_jbot: Save executed orders into database
NOTE: to apply this patch, you need to either:

1. Change DB name, source config again, run trade_jbot so it
   init the table for you. Then, sqlite3 $DB, and run
   >> .read alec/scripts/db.sql
   to create the views.

2. Create table yourself by sqlite3 $DB, and run
   >> CREATE TABLE old_executed_orders (id);
   to create a new table.

Bitfinex v1 rest api is very unreliable.
Either query the order by /orders or /order API,
An order can have its is_live changed from True to False to True to False.
That is, an executed order will be identified to be executed multiple times.
This leads to multiple orders for one pair.

To workaround this, save the executed order ID into db.
We already have a executed_orders table in db.
In order to keep it simple for user, add a new table old_executed_orders
for this usage.
If a newly detected executed order has its ID in old_executed_orders table,
just ignore it. Do not record it again, nor create new orders for it.