simulate_edbot: Add initial version
Based on simulate.cc, add struct and functions for positions and exchange.
Assume bot only creates margin market order.
Add simple strategy to determine next move and backoff time.
It can be improved in the future.

Prepare the data, and modify run.sh to run it.
There are some configs to set:

1. -a to set amount of each order.
2. -p to set the ratio at which we create trailing stop order to protect profit.
   This simulator does not support trailing stop order, so we monitor the price
   periodically and create market order to close the position.
3. -s to set the diff in terms of ratio for trailing stop.
4. -l to set the ratio for stop loss. It should be a negative value.
5. -w to set initial backoff time in seconds. When the guess is wrong, the backoff time will be doubled.
6. -m to set max backoff time.

e.g.

In run.sh, we run
./simulate_edbot -a 10 -p 0.1 -s 0.01 -l -0.2 -w 60 -m 6000
... (a list of data files for each day)

This means, run on ETHUSD data for these days, with order 1 ETH.

- When we guess the right direction:

Say, the position has positive margin ratio at 10%, create a trailing stop order with diff 1%.
That is, when the positive margin ratio goes to 11%, the trailing stop is increased to 11% - 1% = 10% accordingly
so we can have more profit.
When the positive margin ratio is dropped to below 10%, create market order to close the position to take profit.
Reset backoff time after a win.

- When we guess the wrong direction:

Say, the position hit negative 20%. Close the position to stop loss.
Wait 60 seconds to create an opposite order for new position..
If we guess wrong again, double the backoff time until it is above 6000 seconds.