slack_daemon_jbot: Add commands to suspend/resume some targets
We may want to suspend a target when the price is too high.
We may want to resume a target when the price is low enough.

Add the commands to support such requests.

E.g.:
To suspend ethusd:
@trader suspend ethusd

To resume ethusd:
@trader resume ethusd

Also, simplify the command name for disable/enable jbot.

To enable jbot:
@trader enable

To disable jbot:
@trader disable