simulate_edbot: Add life time flow
If there is no new highest ratio for some time since last highest ratio time,
close the position.
If there is a new highest ratio, update the highest ratio and last highest
ratio time.

The backoff mechanism is the same.

E.g.
./simulate_edbot -a 10 -o 86400 -w 60 -m 6000

That is, use life time flow and set it to 86400.
The backoff time is 60, and the max backoff time is 6000.